modbus2mqtt/requirements.txt
minimalmodbus
paho-mqtt==1.6.1

